refactor(tools): extract recipes module
- get_html_recipes serves documents too; static guide tools misplaced in work_items
- Move both recipe tools and gallery models to new recipes.py domain pair
