+++ src/mcp_server_polarion/models/__init__.py
@@ -36,15 +36,17 @@ from mcp_server_polarion.models.links import (
     WorkItemLinkUpdateSpec,
 )
 from mcp_server_polarion.models.projects import ProjectSummary
+from mcp_server_polarion.models.recipes import (
+    HtmlRecipeGallery,
+    SqlRecipeGallery,
+)
 from mcp_server_polarion.models.test_runs import (
     TestRunCreateSpec,
     TestRunsCreateResult,
     TestRunSummary,
 )
 from mcp_server_polarion.models.work_items import (
-    HtmlRecipeGallery,
     Hyperlink,
-    SqlRecipeGallery,
     WorkItemCreateSpec,
     WorkItemDetail,
     WorkItemMoveResult,
