+++ src/mcp_server_polarion/tools/__init__.py
@@ -11,6 +11,7 @@ import mcp_server_polarion.tools.enum
 import mcp_server_polarion.tools.links
 import mcp_server_polarion.tools.moves
 import mcp_server_polarion.tools.projects
+import mcp_server_polarion.tools.recipes
 import mcp_server_polarion.tools.test_runs
 import mcp_server_polarion.tools.work_items  # noqa: F401
