+++ src/mcp_server_polarion/tools/work_items.py
@@ -1,10 +1,9 @@
-"""Work item tools — query, create, update, and SQL recipes."""
+"""Work item tools — query, create, and update."""
 
 from __future__ import annotations
 
 import logging
-from importlib import resources
-from typing import Final, cast
+from typing import cast
 from urllib.parse import urlencode
 
 from fastmcp import Context
@@ -16,10 +15,8 @@ from mcp_server_polarion.core.exceptions import (
     PolarionNotFoundError,
 )
 from mcp_server_polarion.models import (
-    HtmlRecipeGallery,
     JsonValue,
     PaginatedResult,
-    SqlRecipeGallery,
     WorkItemCreateSpec,
     WorkItemDetail,
     WorkItemRead,
@@ -208,19 +205,6 @@ def _update_query_params(
     if change_type_to:
         params["changeTypeTo"] = change_type_to
     return params
-
-
-_SQL_QUERY_RECIPES: Final[str] = (
-    resources.files("mcp_server_polarion.tools")
-    .joinpath("guides", "sql_query_recipes.md")
-    .read_text(encoding="utf-8")
-)
-
-_HTML_RECIPES: Final[str] = (
-    resources.files("mcp_server_polarion.tools")
-    .joinpath("guides", "html_recipes.md")
-    .read_text(encoding="utf-8")
-)
 
 
 @mcp.tool(
@@ -477,36 +461,6 @@ async def update_work_items(  # noqa: PLR0913
         work_item_ids=[spec.work_item_id for spec in items],
         payload_preview=None,
     )
-
-
-@mcp.tool(
-    tags={"read"},
-    annotations={"readOnlyHint": True},
-)
-async def get_sql_query_recipes() -> SqlRecipeGallery:
-    """Fetch copy-paste SQL recipes for the list_work_items SQL:(...) prefix.
-
-    Call before writing any SQL query (document scope, custom-field,
-    traceability); adapt a recipe instead of hand-writing joins. Includes the
-    table schema.
-    """
-    return SqlRecipeGallery(recipes=_SQL_QUERY_RECIPES)
-
-
-@mcp.tool(
-    tags={"read"},
-    annotations={"readOnlyHint": True},
-)
-async def get_html_recipes() -> HtmlRecipeGallery:
-    """Fetch the required HTML templates for tables, captions, links, and
-    widgets written via update_work_items / update_document.
-
-    Any new <table>, numbered caption, work-item / cross-reference / wiki-page
-    link, or TOC / Table-of-Figures widget must be adapted from these
-    templates — plain hand-written markup renders unstyled and breaks
-    numbering. Also covers macro-id and metadata-scope caveats.
-    """
-    return HtmlRecipeGallery(recipes=_HTML_RECIPES)
 
 
 @mcp.tool(
