+++ tests/mcp_server_polarion/tools/test_work_items.py
@@ -15,7 +15,6 @@ from mcp_server_polarion.core.exceptions import (
     PolarionNotFoundError,
 )
 from mcp_server_polarion.models import (
-    HtmlRecipeGallery,
     Hyperlink,
     PaginatedResult,
     WorkItemCreateSpec,
@@ -34,16 +33,10 @@ from mcp_server_polarion.tools.work_items import (
     _build_update_work_items_payload,
     _build_work_item_resource,
     create_work_items,
-    get_html_recipes,
     get_work_item,
     list_work_items,
     read_work_item,
     update_work_items,
-)
-from mcp_server_polarion.utils.html import (
-    _POLARION_TABLE_STYLE,
-    _POLARION_TD_STYLE,
-    _POLARION_TH_STYLE,
 )
 
 
@@ -607,45 +600,6 @@ class TestCreateWorkItemsHappyPath:
         _, kwargs = mock_client.post.call_args
         attributes = kwargs["json"]["data"][0]["attributes"]
         assert attributes["verification_criteria"] == rich
-
-
-class TestGetHtmlRecipes:
-    """``get_html_recipes`` serves raw-HTML templates for update tools."""
-
-    async def test_returns_gallery_with_core_templates(self) -> None:
-        result = await get_html_recipes()
-        assert isinstance(result, HtmlRecipeGallery)
-        assert "polarion-Document-table" in result.recipes
-        assert "polarion-rte-caption-paragraph" in result.recipes
-        assert 'data-sequence="Table"' in result.recipes
-        assert 'data-sequence="Figure"' in result.recipes
-
-    async def test_returns_link_and_widget_templates(self) -> None:
-        result = await get_html_recipes()
-        for marker in (
-            'data-type="workItem"',
-            'data-type="crossReference"',
-            'data-type="richPage"',
-            "polarion_wiki macro name=toc",
-            "polarion_wiki macro name=tof",
-            "polarion_wiki macro name=page_break",
-        ):
-            assert marker in result.recipes
-
-    async def test_recipes_warn_about_scope_and_macro_ids(self) -> None:
-        result = await get_html_recipes()
-        assert "custom field" in result.recipes.lower()
-        assert "polarion_wiki" in result.recipes
-
-    async def test_recipes_stay_in_sync_with_pipeline_constants(self) -> None:
-        # Guide must quote the exact styles polarionify_html injects.
-        result = await get_html_recipes()
-        for constant in (
-            _POLARION_TABLE_STYLE,
-            _POLARION_TH_STYLE,
-            _POLARION_TD_STYLE,
-        ):
-            assert constant in result.recipes
 
 
 class TestCreateWorkItemsErrorMapping:
